backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.17
httpx==0.28.0
anthropic==0.40.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.0.0
